third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-name-multicol-003-expected.txt

This is a testharness.js-based test.
[FAIL] .target 1
  assert_equals: \n<div class="target" data-expected-width="10" data-expected-height="100"></div>\nwidth expected 10 but got 40
[FAIL] .target 2
  assert_equals: \n<div class="target" data-expected-width="10" data-expected-height="100"></div>\nwidth expected 10 but got 70
Harness: the test ran to completion.
